fix(VsTable): move type guard functions from types.ts
