+++ packages/vlossom/src/components/vs-table/composables/table-expand-composable.ts
@@ -1,5 +1,6 @@
 import { computed, reactive, type Ref } from 'vue';
-import { isVsTableBodyRow, getRowId, getRowItem, type VsTableItem, type VsTableCell } from './../types';
+import { type VsTableItem, type VsTableCell } from './../types';
+import { isVsTableBodyRow, getRowId, getRowItem } from './../models/table-model';
 
 export function useTableExpand(
     expandable: Ref<(item: VsTableItem, index?: number, items?: VsTableItem[]) => boolean>,
