+++ packages/vlossom/src/components/vs-table/models/table-cell-builder.ts
@@ -1,6 +1,6 @@
 import { objectUtil } from '@/utils';
 import type { VsTableBodyCell, VsTableCell, VsTableColumnDef, VsTableHeaderCell, VsTableItem } from './../types';
-import { isVsTableColumnDefArray } from './../types';
+import { isVsTableColumnDefArray } from './table-model';
 import {
     NoColumnDefCellStrategy,
     ObjectColumnDefCellStrategy,
