+++ packages/vlossom/src/components/vs-table/types.ts
@@ -93,27 +93,3 @@ export interface VsTableBodyCell<I = VsTableItem> extends VsTableCell<I> {
     tag: 'td';
     item: I;
 }
-
-export function isVsTableColumnDef(value: unknown): value is VsTableColumnDef {
-    return typeof value === 'object' && value !== null && 'key' in value && 'label' in value;
-}
-
-export function isVsTableColumnDefArray(value: unknown): value is VsTableColumnDef[] {
-    return Array.isArray(value) && value.length > 0 && value.every(isVsTableColumnDef);
-}
-
-export function isVsTableBodyRow(row: VsTableCell[]): row is VsTableBodyCell[] {
-    return row[0]?.tag === 'td';
-}
-
-export function getRowItem(row: VsTableBodyCell[]): VsTableItem {
-    const anyCell = row[0];
-    if (!anyCell) {
-        return {};
-    }
-    return anyCell.item;
-}
-
-export function getRowId(row: VsTableBodyCell[]): string | undefined {
-    return row[0]?.id ?? undefined;
-}
